PDFBOX-4892: move fontbox-related test to fontbox; use newer JUnit classes
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1895556 13f79535-47bb-0310-9956-ffa450edef68
